coapcore
Struct OscoreEdhocHandler

pub struct OscoreEdhocHandler<H: Handler, Crypto: Crypto, CryptoFactory: Fn() -> Crypto, SSC: ServerSecurityConfig, RNG: RngCore + CryptoRng, TP: TimeProvider> { /* private fields */ }

A CoAP handler wrapping inner resources, and adding EDHOC and OSCORE and ACE support.

While the ACE (authz-info) and EDHOC parts could be implemented as a handler that is to be added into the tree, the OSCORE part needs to wrap the inner handler anyway, and EDHOC and OSCORE are intertwined rather strongly in processing the EDHOC option.

Implementations

impl<H: Handler, Crypto: Crypto, CryptoFactory: Fn() -> Crypto, SSC: ServerSecurityConfig, RNG: RngCore + CryptoRng, TP: TimeProvider> OscoreEdhocHandler<H, Crypto, CryptoFactory, SSC, RNG, TP>

pub fn new( inner: H, authorities: SSC, crypto_factory: CryptoFactory, rng: RNG, time: TP, ) -> Self

Creates a new CoAP server implementation (a [Handler][coap_handler::Handler]).

The time provider is used to evaluate any time limited tokens leniently; choosing a “bad” time source here (in particular crate::time::TimeUnknown leads to acceptance of expired tokens.