coapcore::scope

Trait Scope

pub trait Scope: Sized + Debug {
    // Required method
    fn request_is_allowed<M: ReadableMessage>(&self, request: &M) -> bool;
}

A data item representing the server access policy as evaluated for a particular security context.

Required Methods

fn request_is_allowed<M: ReadableMessage>(&self, request: &M) -> bool

Returns true if a request may be performed by the bound security context.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.